EM Algorithm Wrapper Function for Deconvolution of Methylation Data

This function wraps around the core EM deconvolution logic, handling array or sequencing data and processing the inputs. It supports parallel processing and allows for different types of input data (arrays or sequencing data) to be processed accordingly.

EM_wrap(
  type,
  bulk_beta = NULL,
  ref_beta = NULL,
  bulk_meth = NULL,
  bulk_cov = NULL,
  ref_meth = NULL,
  ref_cov = NULL,
  bulk_rna = NULL,
  ref_rna = NULL,
  num_iterations = 150,
  bulk_fac,
  mapping = NULL,
  stopping = 1e-04,
  stopping_pi = 5e-04,
  cell_size,
  omega = NULL,
  parallel = TRUE,
  ncore = 90
)

Arguments

type

Specifies the type of DNAm data being used ('array' or 'seq').

bulk_beta

Bulk methylation beta values (if applicable).

ref_beta

Reference DNAm beta values.

bulk_meth

Bulk DNAm methylated counts data.

bulk_cov

Bulk DNAM coverage data.

ref_meth

Reference DNAm methylated counts data.

ref_cov

Reference coverage data.

bulk_rna

Bulk RNA data.

ref_rna

Reference RNA data.

num_iterations

Number of EM iterations to perform.

bulk_fac

Bulk factor used in processing.

mapping

Optional mapping parameter.

stopping

Stopping criterion for EM convergence.

stopping_pi

Stopping criterion for pi convergence.

cell_size

Cell size of RNA deconvolution.

omega

Parameter to choose between DNAm (1) part and RNA (0) part.

parallel

Logical flag indicating whether to enable parallel computation. If TRUE, parallel computation will be used.

ncore

Number of cores for parallel processing.

Value

Returns a list of estimated deconvolution results.

Examples

# Example usage for sequencing data:
result <- EM_wrap("seq", bulk_beta = bulk_beta_matrix, ref_beta = ref_beta_matrix, parallel = TRUE)
#> Error in EM_wrap("seq", bulk_beta = bulk_beta_matrix, ref_beta = ref_beta_matrix,     parallel = TRUE): object 'ref_beta_matrix' not found